requirements.txt
asgiref==3.8.0
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.2
django-import-export==4.0.0
django-property-filter==1.2.0
djangorestframework==3.15.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
gunicorn==22.0.0
inflection==0.5.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.6.0
typing_extensions==4.11.0
unicodecsv==0.14.1
uritemplate==4.1.1
xlrd==2.0.1
xlwt==1.3.0
